Improved ContentType parser, EncodedWords parser, QuotedPrintable parser, Base32/64 decoders. Improved HeaderLineReader. Improved junit code coverage in common package. Implemented OutputStream wrapper for GZip writer. Added support for GZip compressed WARC writing.

+++ jwat-common/src/main/java/org/jwat/common/HeaderLine.java
@@ -26,17 +26,17 @@ package org.jwat.common;
  */
 public class HeaderLine {
 
-	/** Invalid line, for reporting the raw data read. */
-	public static final int HLT_RAW = 0;
-	/** Normal line parsed. */
-	public static final int HLT_LINE = 1;
-	/** Header line parsed. */
-	public static final int HLT_HEADERLINE = 2;
+    /** Invalid line, for reporting the raw data read. */
+    public static final byte HLT_RAW = 0;
+    /** Normal line parsed. */
+    public static final byte HLT_LINE = 1;
+    /** Header line parsed. */
+    public static final byte HLT_HEADERLINE = 2;
 
-	/** Type of the parsed line. */
-	public byte type = HLT_RAW;;
+    /** Type of the parsed line. */
+    public byte type = HLT_RAW;;
 
-	/** Header name. */
+    /** Header name. */
     public String name;
 
     /** Header value. */
